requirements.txt
streamlit==1.39.0
openai==1.51.0
requests==2.31.0
python-dotenv==1.0.0
gspread==6.1.4
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1